[FEAT] 인기게시글 알람 1차 구현

+++ iz/src/main/java/com/izikgram/alarm/service/SseEmitterService.java
@@ -43,7 +43,21 @@ public class SseEmitterService {
             try {
                 // 3️⃣ SSE를 통해 "새로운 댓글이 달렸습니다!" 이벤트 전송
                 sseEmitter.send(SseEmitter.event().name("message").data(content));
-            } catch (Exception e) {
+            } catch (IOException e) {
+                // 4️⃣ SSE 연결이 끊어졌다면 해당 사용자 제거
+                sseEmitterRepository.remove(member_id);
+            }
+        }
+    }
+
+    public void popularSend(String member_id, String content) {
+        // 2️⃣ 작성자가 SSE에 연결되어 있으면 알림 전송
+        if (sseEmitterRepository.containsKey(member_id)) {
+            SseEmitter sseEmitter = sseEmitterRepository.findByMemberId(member_id);
+            try {
+                // 3️⃣ SSE를 통해 "새로운 댓글이 달렸습니다!" 이벤트 전송
+                sseEmitter.send(SseEmitter.event().name("popular-message").data(content));
+            } catch (IOException e) {
                 // 4️⃣ SSE 연결이 끊어졌다면 해당 사용자 제거
                 sseEmitterRepository.remove(member_id);
             }
